Some refactoring for better generalization
